refactor: remove dca runtime tools and modules
