+++ kis_websocket_monitor.py
@@ -2,8 +2,7 @@
 """
 KIS WebSocket Monitor - Main Entry Point
 
-KIS 체결 WebSocket을 모니터링하여 체결 이벤트를 Redis pub/sub으로 발행하고
-DCA 플랜 체결 단계를 자동 업데이트합니다.
+KIS 체결 WebSocket을 모니터링하여 체결 이벤트를 Redis pub/sub으로 발행합니다.
 
 사용법:
     python kis_websocket_monitor.py
@@ -15,19 +14,10 @@ import asyncio
 import logging
 import signal
 import sys
-from decimal import Decimal
 from typing import Any
-
-from sqlalchemy.ext.asyncio import (
-    AsyncEngine,
-    AsyncSession,
-    async_sessionmaker,
-    create_async_engine,
-)
 
 from app.core.config import settings
 from app.monitoring.sentry import capture_exception, init_sentry
-from app.services.dca_service import DcaService
 from app.services.execution_event import (
     close_redis as close_execution_redis,
 )
@@ -43,13 +33,11 @@ class KISWebSocketMonitor:
     """
     KIS WebSocket 모니터
 
-    WebSocket 연결, 체결 처리, DCA 통합을 담당합니다.
+    WebSocket 연결, 체결 처리, 이벤트 발행을 담당합니다.
     """
 
     def __init__(self):
-        self.dca_service: DcaService | None = None
         self.websocket_client: KISExecutionWebSocket | None = None
-        self._db_engine: AsyncEngine | None = None
         self.is_running = False
         self._setup_signal_handlers()
 
@@ -77,30 +65,6 @@ class KISWebSocketMonitor:
         )
         self.is_running = False
 
-    def _initialize_db(self) -> async_sessionmaker[AsyncSession]:
-        """
-        데이터베이스 엔진 초기화
-
-        Returns:
-            async_sessionmaker[AsyncSession]: 비동기 DB 세션 팩토리
-        """
-        self._db_engine = create_async_engine(settings.DATABASE_URL, echo=False)
-        return async_sessionmaker(
-            self._db_engine,
-            expire_on_commit=False,
-            class_=AsyncSession,
-        )
-
-    async def _initialize_dca_service(self, db_session: AsyncSession):
-        """
-        DCA 서비스 초기화
-
-        Args:
-            db_session: DB 세션
-        """
-        self.dca_service = DcaService(db_session)
-        logger.info("DCA service initialized")
-
     async def _initialize_websocket(self):
         """
         KIS WebSocket 클라이언트 초기화
@@ -117,68 +81,13 @@ class KISWebSocketMonitor:
         """
         체결 이벤트 수신 처리
 
-        DCA 체결 단계 업데이트 및 Redis 이벤트 발행을 수행합니다.
-        order_id가 있는 경우에만 DCA 통합을 시도합니다.
+        체결 이벤트를 수신하면 Redis 이벤트를 발행합니다.
 
         Args:
             event: 체결 이벤트 데이터
                 symbol, side, order_id, filled_price, filled_qty, exec_time, timestamp, market
         """
-        order_id = event.get("order_id")
-        if not order_id:
-            logger.debug("Execution event without order_id, skipping DCA update")
-            await self._publish_execution_event(event)
-            return
-
-        try:
-            await self._update_dca_step(order_id, event)
-        except Exception as e:
-            logger.warning(f"DCA step update failed: {e}, continuing...")
-
         await self._publish_execution_event(event)
-
-    async def _update_dca_step(self, order_id: str, event: dict[str, Any]) -> None:
-        """
-        DCA 체결 단계 업데이트
-
-        order_id로 DCA step를 찾아 FILLED로 표시합니다.
-        다음 pending step가 있으면 이벤트에 추가합니다.
-
-        Args:
-            order_id: 주문 ID
-            event: 체결 이벤트 데이터
-        """
-        if not self.dca_service:
-            return
-
-        step = await self.dca_service.find_step_by_order_id(order_id)
-        if not step:
-            logger.debug(f"No DCA step found for order_id={order_id}")
-            return
-
-        await self.dca_service.mark_step_filled(
-            step_id=step.id,
-            filled_price=Decimal(str(event.get("filled_price", 0))),
-            filled_qty=Decimal(str(event.get("filled_qty", 0))),
-        )
-
-        logger.info(
-            f"DCA step marked as filled: step_id={step.id}, "
-            f"order_id={order_id}, plan_id={step.plan_id}"
-        )
-
-        next_step = await self.dca_service.get_next_pending_step(step.plan_id)
-        if next_step:
-            event["dca_next_step"] = {
-                "plan_id": next_step.plan_id,
-                "step_number": next_step.step_number,
-                "target_price": float(next_step.target_price),
-                "target_quantity": str(next_step.target_quantity),
-            }
-            logger.info(
-                f"Next pending step added to event: plan_id={next_step.plan_id}, "
-                f"step_number={next_step.step_number}"
-            )
 
     async def _publish_execution_event(self, event: dict[str, Any]) -> None:
         """
@@ -201,17 +110,13 @@ class KISWebSocketMonitor:
         logger.info("Starting KIS WebSocket monitor...")
         self.is_running = True
 
-        async_session_maker = self._initialize_db()
+        await self._initialize_websocket()
+        if self.websocket_client is None:
+            raise RuntimeError("KIS WebSocket client initialization failed")
 
-        async with async_session_maker() as db_session:
-            await self._initialize_dca_service(db_session)
-            await self._initialize_websocket()
-            if self.websocket_client is None:
-                raise RuntimeError("KIS WebSocket client initialization failed")
+        await self.websocket_client.connect_and_subscribe()
 
-            await self.websocket_client.connect_and_subscribe()
-
-            await self.websocket_client.listen()
+        await self.websocket_client.listen()
 
     async def stop(self):
         """
@@ -224,10 +129,6 @@ class KISWebSocketMonitor:
 
         if self.websocket_client:
             await self.websocket_client.stop()
-
-        if self._db_engine is not None:
-            await self._db_engine.dispose()
-            self._db_engine = None
 
         await close_execution_redis()
