+++ tests/test_mcp_server_tools.py
@@ -1,6 +1,5 @@
 import logging
 from datetime import date
-from decimal import Decimal
 from unittest.mock import AsyncMock
 
 import httpx
@@ -32,11 +31,9 @@ from app.mcp_server.tooling import (
     shared,
 )
 from app.mcp_server.tooling.registry import register_all_tools
-from app.models.dca_plan import DcaPlan, DcaPlanStatus, DcaPlanStep, DcaStepStatus
 from app.services import naver_finance
 from app.services import upbit as upbit_service
 from app.services import yahoo as yahoo_service
-from app.services.dca_service import DcaService
 
 # from app.mcp_server.tick_size import adjust_tick_size_kr  # TODO: Remove if not needed
 
@@ -70,6 +67,25 @@ def build_tools() -> dict[str, object]:
     mcp = DummyMCP()
     register_all_tools(mcp)
     return mcp.tools
+
+
+@pytest.mark.asyncio
+async def test_dca_tools_removed_from_build_tools() -> None:
+    tools = build_tools()
+
+    assert "create_dca_plan" not in tools
+    assert "get_dca_status" not in tools
+
+
+def test_models_package_no_longer_exports_dca() -> None:
+    import app.models as models
+
+    assert not hasattr(models, "DcaPlan")
+    assert not hasattr(models, "DcaPlanStep")
+
+
+def test_compute_dca_price_levels_helper_removed() -> None:
+    assert not hasattr(market_data_indicators, "_compute_dca_price_levels")
 
 
 _PATCH_MODULES = (
@@ -6237,1137 +6253,6 @@ class TestComputeRsiWeights:
         # All weights equal
         expected_weight = 1.0 / 3
         assert all(abs(w - expected_weight) < 0.001 for w in result)
-
-
-@pytest.mark.unit
-class TestComputeDcaPriceLevels:
-    """Tests for _compute_dca_price_levels helper function."""
-
-    def test_support_strategy_with_sufficient_supports(self):
-        """Support strategy with enough support levels."""
-        current_price = 100000.0
-        supports = [
-            {"price": 95000.0, "source": "fib_23.6"},
-            {"price": 90000.0, "source": "fib_38.2"},
-            {"price": 85000.0, "source": "fib_50.0"},
-            {"price": 80000.0, "source": "fib_61.8"},
-        ]
-
-        result = market_data_indicators._compute_dca_price_levels(
-            "support", 3, current_price, supports
-        )
-
-        assert len(result) == 3
-        # Should use closest 3 supports
-        assert result[0]["price"] == 95000.0
-        assert result[1]["price"] == 90000.0
-        assert result[2]["price"] == 85000.0
-
-    def test_support_strategy_with_fewer_supports(self):
-        """Support strategy with fewer than splits supports."""
-        current_price = 100000.0
-        supports = [
-            {"price": 90000.0, "source": "fib_38.2"},
-            {"price": 80000.0, "source": "fib_61.8"},
-        ]
-
-        result = market_data_indicators._compute_dca_price_levels(
-            "support", 4, current_price, supports
-        )
-
-        assert len(result) == 4
-        # Should interpolate between supports
-        # First levels near supports, last levels interpolated
-        assert all(level["source"] == "support" for level in result)
-
-    def test_support_strategy_with_no_supports(self):
-        """Support strategy with no support levels."""
-        current_price = 100000.0
-        supports = []
-
-        result = market_data_indicators._compute_dca_price_levels(
-            "support", 3, current_price, supports
-        )
-
-        assert len(result) == 3
-        # Should create synthetic levels: -2%, -4%, -6%
-        assert result[0]["source"] == "synthetic"
-        assert abs(result[0]["price"] / current_price - 0.98) < 0.01
-        assert abs(result[1]["price"] / current_price - 0.96) < 0.01
-        assert abs(result[2]["price"] / current_price - 0.94) < 0.01
-
-    def test_equal_strategy(self):
-        """Equal strategy with supports."""
-        current_price = 100000.0
-        supports = [
-            {"price": 85000.0, "source": "fib_50.0"},
-        ]
-
-        result = market_data_indicators._compute_dca_price_levels(
-            "equal", 3, current_price, supports
-        )
-
-        assert len(result) == 3
-        assert result[0]["source"] == "equal_spaced"
-        # Should space from current_price to min support
-        assert result[0]["price"] < current_price
-        assert result[-1]["price"] == 85000.0
-
-    def test_equal_strategy_without_supports(self):
-        """Equal strategy without supports (goes to -10%)."""
-        current_price = 100000.0
-        supports = []
-
-        result = market_data_indicators._compute_dca_price_levels(
-            "equal", 3, current_price, supports
-        )
-
-        assert len(result) == 3
-        assert result[-1]["price"] == 90000.0  # -10%
-
-    def test_aggressive_strategy(self):
-        """Aggressive strategy: first buy at current - 0.5%, rest support."""
-        current_price = 100000.0
-        supports = [
-            {"price": 90000.0, "source": "fib_61.8"},
-            {"price": 80000.0, "source": "fib_61.8"},
-        ]
-
-        result = market_data_indicators._compute_dca_price_levels(
-            "aggressive", 3, current_price, supports
-        )
-
-        assert len(result) == 3
-        assert result[0]["source"] == "aggressive_first"
-        assert abs(result[0]["price"] / current_price - 0.995) < 0.001
-        # Rest 2 levels from support strategy
-        assert result[1]["source"] in ["support", "equal_spaced", "interpolated"]
-        assert result[2]["source"] in ["support", "equal_spaced", "interpolated"]
-
-    def test_invalid_strategy_raises_error(self):
-        """Invalid strategy raises ValueError."""
-        current_price = 100000.0
-        supports = []
-
-        with pytest.raises(ValueError, match="Invalid strategy"):
-            market_data_indicators._compute_dca_price_levels(
-                "invalid", 3, current_price, supports
-            )
-
-
-@pytest.mark.asyncio
-class TestCreateDcaPlan:
-    """Tests for create_dca_plan MCP tool."""
-
-    async def test_create_dca_plan_market_hint_regression(self, monkeypatch):
-        tools = build_tools()
-
-        sr_calls: list[tuple[str, str | None]] = []
-        indicator_calls: list[tuple[str, list[str], str | None]] = []
-
-        async def mock_sr(symbol: str, market: str | None):
-            sr_calls.append((symbol, market))
-            return {
-                "symbol": symbol,
-                "current_price": 80000.0,
-                "supports": [{"price": 76000.0, "source": "fib_23.6"}],
-                "resistances": [],
-            }
-
-        async def mock_indicators(
-            symbol: str, indicators: list[str], market: str | None
-        ):
-            indicator_calls.append((symbol, indicators, market))
-            return {
-                "symbol": symbol,
-                "price": 80000.0,
-                "indicators": {"rsi": {"14": 42.0}},
-            }
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        created_plan = DcaPlan(
-            id=101,
-            user_id=1,
-            symbol="005930",
-            market="equity_kr",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        created_plan.total_amount = Decimal("900000")
-
-        async def mock_create_plan(self, **kwargs):
-            return created_plan
-
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return created_plan
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        with pytest.raises(
-            ValueError, match="korean equity symbols must be 6 alphanumeric characters"
-        ):
-            await tools["create_dca_plan"](
-                symbol="AAPL",
-                market="kr",
-                total_amount=900000.0,
-                splits=3,
-                strategy="support",
-                dry_run=True,
-            )
-
-        assert sr_calls == []
-        assert indicator_calls == []
-
-        result = await tools["create_dca_plan"](
-            symbol="005930",
-            market="kr",
-            total_amount=900000.0,
-            splits=3,
-            strategy="support",
-            dry_run=True,
-        )
-
-        assert result["success"] is True
-        assert sr_calls[-1] == ("005930", "kr")
-        assert indicator_calls[-1][0] == "005930"
-        assert indicator_calls[-1][1] == ["rsi"]
-        assert indicator_calls[-1][2] == "kr"
-
-    async def test_support_strategy_dry_run_crypto(self, monkeypatch):
-        """Support strategy with dry_run=True for crypto."""
-        tools = build_tools()
-
-        # Mock get_support_resistance
-        mock_sr_result = {
-            "symbol": "KRW-BTC",
-            "current_price": 100000000.0,
-            "supports": [
-                {"price": 95000000.0, "source": "fib_23.6"},
-                {"price": 90000000.0, "source": "fib_38.2"},
-            ],
-            "resistances": [],
-        }
-
-        async def mock_sr(_symbol, _market):
-            return mock_sr_result
-
-        # Mock get_indicators (RSI < 30 = oversold → front_heavy)
-        mock_indicator_result = {
-            "symbol": "KRW-BTC",
-            "price": 100000000.0,
-            "indicators": {
-                "rsi": {"14": 25.0},
-            },
-        }
-
-        async def mock_indicators(_symbol, _indicators, _market):
-            return mock_indicator_result
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        # Mock DB session and DcaService.create_plan so persistence succeeds for dry_run
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        created_plan = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        created_plan.total_amount = Decimal("200000")
-
-        async def mock_create_plan(self, **kwargs):
-            return created_plan
-
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return created_plan
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        result = await tools["create_dca_plan"](
-            symbol="KRW-BTC",
-            total_amount=200000.0,
-            splits=3,
-            strategy="support",
-            dry_run=True,
-        )
-
-        assert result["success"] is True
-        assert result["dry_run"] is True
-        assert "plans" in result
-        assert "summary" in result
-        assert len(result["plans"]) == 3
-
-        # Check summary
-        summary = result["summary"]
-        assert summary["symbol"] == "KRW-BTC"
-        assert summary["current_price"] == 100000000.0
-        assert summary["rsi"] == 25.0
-        assert "rsi_14" not in summary
-        assert summary["strategy"] == "support"
-        assert summary["total_amount"] == 200000.0
-        assert summary["weight_mode"] == "front_heavy"  # RSI < 30
-
-        # Check plans
-        plans = result["plans"]
-        assert plans[0]["step"] == 1
-        assert plans[0]["source"] in ["support", "fib_23.6"]
-        assert abs(sum(p["amount"] for p in plans) - 200000.0) < 1.0
-
-    async def test_equal_strategy_dry_run_kr_equity(self, monkeypatch):
-        """Equal strategy with dry_run=True for KR equity."""
-        tools = build_tools()
-
-        mock_sr_result = {
-            "symbol": "005930",
-            "current_price": 80000.0,
-            "supports": [
-                {"price": 75000.0, "source": "fib_23.6"},
-            ],
-            "resistances": [],
-        }
-
-        async def mock_sr(_symbol, _market):
-            return mock_sr_result
-
-        mock_indicator_result = {
-            "symbol": "005930",
-            "price": 80000.0,
-            "indicators": {
-                "rsi": {"14": 45.0},
-            },
-        }
-
-        async def mock_indicators(_symbol, _indicators, _market):
-            return mock_indicator_result
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        # Mock DB session and DcaService.create_plan so persistence succeeds for dry_run
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        created_plan = DcaPlan(
-            id=2,
-            user_id=1,
-            symbol="005930",
-            market="equity_kr",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        created_plan.total_amount = Decimal("1000000")
-
-        async def mock_create_plan(self, **kwargs):
-            return created_plan
-
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return created_plan
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        result = await tools["create_dca_plan"](
-            symbol="005930",
-            total_amount=1000000.0,
-            splits=4,
-            strategy="equal",
-            dry_run=True,
-        )
-
-        assert result["success"] is True
-        assert result["dry_run"] is True
-        assert len(result["plans"]) == 4
-
-        # KR equity uses integer quantities
-        assert all(isinstance(p["quantity"], int) for p in result["plans"])
-        # Equal weights: neutral RSI (30-50)
-        assert result["summary"]["weight_mode"] == "equal"
-
-    async def test_create_dca_plan_kr_valid_ticks_omit_tick_adjusted_metadata(
-        self, monkeypatch
-    ):
-        """KR DCA plan should omit tick metadata when all prices are already valid ticks."""
-        tools = build_tools()
-
-        mock_sr_result = {
-            "symbol": "012450",
-            "current_price": 1_120_000.0,
-            "supports": [
-                {"price": 1_098_000.0, "source": "fib_23.6"},
-                {"price": 1_096_000.0, "source": "fib_38.2"},
-            ],
-            "resistances": [],
-        }
-
-        async def mock_sr(_symbol, _market):
-            return mock_sr_result
-
-        mock_indicator_result = {
-            "symbol": "012450",
-            "price": 1_120_000.0,
-            "indicators": {"rsi": {"14": 40.0}},
-        }
-
-        async def mock_indicators(_symbol, _indicators, _market):
-            return mock_indicator_result
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        created_plan = DcaPlan(
-            id=3,
-            user_id=1,
-            symbol="012450",
-            market="equity_kr",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        created_plan.total_amount = Decimal("3000000")
-
-        async def mock_create_plan(self, **kwargs):
-            return created_plan
-
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return created_plan
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        result = await tools["create_dca_plan"](
-            symbol="012450",
-            total_amount=3_000_000.0,
-            splits=2,
-            strategy="support",
-            dry_run=True,
-        )
-
-        assert result["success"] is True
-        assert result["dry_run"] is True
-        assert len(result["plans"]) == 2
-        assert result["plans"][0]["price"] == 1_098_000.0
-        assert result["plans"][1]["price"] == 1_096_000.0
-        assert all("tick_adjusted" not in plan for plan in result["plans"])
-        assert all("original_price" not in plan for plan in result["plans"])
-
-    async def test_create_dca_plan_kr_invalid_ticks_include_tick_adjusted_metadata(
-        self, monkeypatch
-    ):
-        """KR DCA plan should include tick metadata only when adjustment is applied."""
-        tools = build_tools()
-
-        mock_sr_result = {
-            "symbol": "012450",
-            "current_price": 1_120_000.0,
-            "supports": [
-                {"price": 1_098_500.0, "source": "fib_23.6"},
-                {"price": 1_096_300.0, "source": "fib_38.2"},
-            ],
-            "resistances": [],
-        }
-
-        async def mock_sr(_symbol, _market):
-            return mock_sr_result
-
-        mock_indicator_result = {
-            "symbol": "012450",
-            "price": 1_120_000.0,
-            "indicators": {"rsi": {"14": 40.0}},
-        }
-
-        async def mock_indicators(_symbol, _indicators, _market):
-            return mock_indicator_result
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        created_plan = DcaPlan(
-            id=4,
-            user_id=1,
-            symbol="012450",
-            market="equity_kr",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        created_plan.total_amount = Decimal("3000000")
-
-        async def mock_create_plan(self, **kwargs):
-            return created_plan
-
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return created_plan
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        result = await tools["create_dca_plan"](
-            symbol="012450",
-            total_amount=3_000_000.0,
-            splits=2,
-            strategy="support",
-            dry_run=True,
-        )
-
-        assert result["success"] is True
-        assert result["dry_run"] is True
-        assert len(result["plans"]) == 2
-        assert result["plans"][0]["price"] == 1_098_000.0
-        assert result["plans"][0]["tick_adjusted"] is True
-        assert result["plans"][0]["original_price"] == 1_098_500.0
-        assert result["plans"][1]["price"] == 1_096_000.0
-        assert result["plans"][1]["tick_adjusted"] is True
-        assert result["plans"][1]["original_price"] == 1_096_300.0
-
-    async def test_create_dca_plan_places_order_and_marks_step_ordered(
-        self, monkeypatch
-    ):
-        """Non-dry-run path should place order and mark first step ordered."""
-        tools = build_tools()
-
-        # Mock market resolution helpers
-        mock_sr_result = {
-            "symbol": "KRW-BTC",
-            "current_price": 100000000.0,
-            "supports": [
-                {"price": 95000000.0, "source": "fib_23.6"},
-                {"price": 90000000.0, "source": "fib_38.2"},
-            ],
-            "resistances": [],
-        }
-
-        async def mock_sr(_symbol, _market):
-            return mock_sr_result
-
-        mock_indicator_result = {
-            "symbol": "KRW-BTC",
-            "price": 100000000.0,
-            "indicators": {
-                "rsi": {"14": 40.0},
-            },
-        }
-
-        async def mock_indicators(_symbol, _indicators, _market):
-            return mock_indicator_result
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        # Mock DB session factory
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        # Prepare a created plan with steps
-        step1 = DcaPlanStep(
-            id=10,
-            plan_id=1,
-            step_number=1,
-            target_price=Decimal("95000000"),
-            target_amount=Decimal("100000"),
-            target_quantity=Decimal("0.001"),
-            status=DcaStepStatus.PENDING,
-        )
-        step2 = DcaPlanStep(
-            id=11,
-            plan_id=1,
-            step_number=2,
-            target_price=Decimal("90000000"),
-            target_amount=Decimal("100000"),
-            target_quantity=Decimal("0.001"),
-            status=DcaStepStatus.PENDING,
-        )
-        created_plan = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[step1, step2],
-        )
-        created_plan.total_amount = Decimal("100000")
-        created_plan.splits = 2
-        created_plan.strategy = "support"
-
-        async def mock_create_plan(self, **kwargs):
-            return created_plan
-
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return created_plan
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        # Mock order placement
-        async def mock_place_order_impl(*args, **kwargs):
-            return {"success": True, "order_id": "ORDER-XYZ"}
-
-        _patch_runtime_attr(monkeypatch, "_place_order_impl", mock_place_order_impl)
-
-        # Mock mark_step_ordered
-        mock_mark_step_ordered = AsyncMock()
-        monkeypatch.setattr(DcaService, "mark_step_ordered", mock_mark_step_ordered)
-
-        # Execute
-        result = await tools["create_dca_plan"](
-            symbol="KRW-BTC",
-            total_amount=200000.0,
-            splits=2,
-            strategy="support",
-            dry_run=False,
-            execute_steps=[1],
-        )
-
-        # Verify result
-        assert result["success"] is True
-        assert result["dry_run"] is False
-        assert result.get("executed") is True
-        assert "plan_id" in result
-        assert result["plan_id"] == 1
-        assert "plans" in result
-        assert "summary" in result
-
-        # Verify order placement -> mark_step_ordered called with correct order_id
-        mock_mark_step_ordered.assert_awaited_once()
-        called_args, _ = mock_mark_step_ordered.await_args
-        # called_args: (step_id, order_id) because AsyncMock is patched on the class
-        assert called_args[0] == step1.id
-        assert called_args[1] == "ORDER-XYZ"
-
-
-@pytest.mark.asyncio
-class TestCreateDcaPlanValidation:
-    """Validation tests for create_dca_plan MCP tool."""
-
-    async def test_invalid_symbol_raises_error(self, monkeypatch):
-        """Empty symbol raises ValueError."""
-        tools = build_tools()
-
-        with pytest.raises(ValueError, match="symbol is required"):
-            await tools["create_dca_plan"](
-                symbol="",
-                total_amount=100000.0,
-                splits=3,
-                dry_run=True,
-            )
-
-    async def test_invalid_splits_raises_error(self, monkeypatch):
-        """splits outside 2-5 range raises ValueError."""
-        tools = build_tools()
-
-        with pytest.raises(ValueError, match="splits must be between 2 and 5"):
-            await tools["create_dca_plan"](
-                symbol="KRW-BTC",
-                total_amount=100000.0,
-                splits=6,
-                dry_run=True,
-            )
-
-
-class TestGetDcaStatus:
-    """Tests for get_dca_status MCP tool."""
-
-    @staticmethod
-    def build_tools():
-        mcp = DummyMCP()
-        register_all_tools(mcp)
-        return mcp.tools
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_by_plan_id(self, monkeypatch):
-        """Test get_dca_status by plan_id (highest priority)."""
-        tools = build_tools()
-
-        # Mock database
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        # Mock DcaService.get_plan
-        plan = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[
-                DcaPlanStep(
-                    id=1,
-                    plan_id=1,
-                    step_number=1,
-                    status=DcaStepStatus.FILLED,
-                ),
-                DcaPlanStep(
-                    id=2,
-                    plan_id=1,
-                    step_number=2,
-                    status=DcaStepStatus.PENDING,
-                ),
-                DcaPlanStep(
-                    id=3,
-                    plan_id=1,
-                    step_number=3,
-                    status=DcaStepStatus.ORDERED,
-                ),
-            ],
-        )
-
-        async def mock_get_plan(self, plan_id, _user_id=None):
-            if plan_id == 1:
-                return plan
-            return None
-
-        monkeypatch.setattr(
-            DcaService,
-            "get_plan",
-            mock_get_plan,
-        )
-
-        # Mock DcaService.get_plans_by_status
-        async def mock_get_plans_by_status(self, **_kwargs):
-            # Accept any keyword arguments
-            return [plan]
-
-        monkeypatch.setattr(
-            DcaService,
-            "get_plans_by_status",
-            mock_get_plans_by_status,
-        )
-
-        # Execute
-        result = await tools["get_dca_status"](plan_id=1)
-
-        # Verify
-        assert result["success"] is True
-        assert "plans" in result
-        assert len(result["plans"]) == 1
-        p = result["plans"][0]
-        assert p["plan_id"] == 1
-        assert p["symbol"] == "KRW-BTC"
-        assert p["status"] == "active"
-        assert "rsi" in p
-        assert "rsi_14" not in p
-        # Progress summary present and counts as expected
-        assert "progress" in p
-        assert p["progress"]["total_steps"] == 3
-        assert p["progress"]["filled"] == 1
-        assert p["progress"]["ordered"] == 1
-        assert p["progress"]["pending"] == 1
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_by_symbol(self, monkeypatch):
-        """Test get_dca_status by symbol."""
-        tools = build_tools()
-
-        plan1 = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        plan2 = DcaPlan(
-            id=2,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            status=DcaPlanStatus.COMPLETED,
-            steps=[],
-        )
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        async def mock_get_plans(self, **kwargs):
-            status = kwargs.get("status")
-            if status == "active":
-                return [plan1]
-            elif status is None:
-                return [plan1, plan2]
-            return []
-
-        monkeypatch.setattr(
-            DcaService,
-            "get_plans_by_status",
-            mock_get_plans,
-        )
-
-        # Execute
-        result = await tools["get_dca_status"](symbol="KRW-BTC", status="active")
-
-        # Verify
-        assert result["success"] is True
-        assert "plans" in result
-        assert len(result["plans"]) == 1
-        p = result["plans"][0]
-        assert p["symbol"] == "KRW-BTC"
-        assert p["status"] == "active"
-        assert "progress" in p
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_by_status_only(self, monkeypatch):
-        """Test get_dca_status by status only."""
-        tools = build_tools()
-
-        plan1 = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-ETH",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        # Mock DcaService.get_plans_by_status
-        async def mock_get_plans(self, **kwargs):
-            return [plan1]
-
-        monkeypatch.setattr(
-            DcaService,
-            "get_plans_by_status",
-            mock_get_plans,
-        )
-
-        # Execute
-        result = await tools["get_dca_status"](status="active")
-
-        # Verify
-        assert result["success"] is True
-        assert "plans" in result
-        assert len(result["plans"]) == 1
-        p = result["plans"][0]
-        assert p["plan_id"] == 1
-        assert p["symbol"] == "KRW-ETH"
-        assert "progress" in p
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_all_status(self, monkeypatch):
-        """Test get_dca_status with status='all'."""
-        tools = build_tools()
-
-        plan1 = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-        plan2 = DcaPlan(
-            id=2,
-            user_id=1,
-            symbol="KRW-ETH",
-            market="crypto",
-            status=DcaPlanStatus.CANCELLED,
-            steps=[],
-        )
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        # Mock DcaService.get_plans_by_status
-        async def mock_get_plans(self, **kwargs):
-            return [plan1, plan2]
-
-        monkeypatch.setattr(
-            DcaService,
-            "get_plans_by_status",
-            mock_get_plans,
-        )
-
-        # Execute
-        result = await tools["get_dca_status"](status="all")
-
-        # Verify
-        assert result["success"] is True
-        assert "plans" in result
-        assert len(result["plans"]) == 2
-        assert any(
-            p["status"] == "active" or p["status"] == "cancelled"
-            for p in result["plans"]
-        )
-        for p in result["plans"]:
-            assert "progress" in p
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_invalid_status(self, monkeypatch):
-        """Test get_dca_status with invalid status value."""
-        tools = build_tools()
-
-        # Execute
-        result = await tools["get_dca_status"](status="invalid")
-
-        # Verify error
-        assert result["success"] is False
-        assert "error" in result
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_limit(self, monkeypatch):
-        """Test get_dca_status with limit parameter."""
-        tools = build_tools()
-
-        plans_list = []
-        for i in range(5):
-            plans_list.append(
-                DcaPlan(
-                    id=i,
-                    user_id=1,
-                    symbol="KRW-BTC",
-                    market="crypto",
-                    status=DcaPlanStatus.ACTIVE,
-                    steps=[],
-                )
-            )
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        # Mock DcaService.get_plans_by_status
-        async def mock_get_plans(self, **kwargs):
-            limit = kwargs.get("limit", 10)
-            return plans_list[:limit]
-
-        monkeypatch.setattr(
-            DcaService,
-            "get_plans_by_status",
-            mock_get_plans,
-        )
-
-        # Execute
-        result = await tools["get_dca_status"](limit=3)
-
-        # Verify
-        assert result["success"] is True
-        assert "plans" in result
-        assert len(result["plans"]) == 3
-
-    @pytest.mark.asyncio
-    async def test_create_dca_plan_mark_step_ordered_with_reloaded_plan(
-        self, monkeypatch
-    ):
-        """Test that mark_step_ordered is called even when created_plan.steps is empty."""
-        tools = build_tools()
-
-        # Mock DB session and plan creation
-        mock_step1 = DcaPlanStep(
-            id=1,
-            plan_id=100,
-            step_number=1,
-            target_amount=Decimal("100000"),
-            target_price=Decimal("100000"),
-            target_quantity=Decimal("0.001"),
-            status=DcaStepStatus.PENDING,
-        )
-        mock_step2 = DcaPlanStep(
-            id=2,
-            plan_id=100,
-            step_number=2,
-            target_amount=Decimal("100000"),
-            target_price=Decimal("95000"),
-            target_quantity=Decimal("0.001"),
-            status=DcaStepStatus.PENDING,
-        )
-
-        mock_plan_reloaded = DcaPlan(
-            id=100,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            total_amount=Decimal("100000"),
-            splits=2,
-            strategy="support",
-            rsi_14=50.0,
-            status=DcaPlanStatus.ACTIVE,
-            steps=[mock_step1, mock_step2],
-        )
-
-        mock_plan_created = DcaPlan(
-            id=100,
-            user_id=1,
-            symbol="KRW-BTC",
-            market="crypto",
-            total_amount=Decimal("100000"),
-            splits=2,
-            strategy="support",
-            rsi_14=50.0,
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        # Mock DcaService.create_plan (returns plan without steps)
-        async def mock_create_plan(self, **kwargs):
-            return mock_plan_created
-
-        # Mock DcaService.get_plan (returns plan with steps)
-        async def mock_get_plan(self, plan_id, user_id=None):
-            return mock_plan_reloaded
-
-        monkeypatch.setattr(DcaService, "create_plan", mock_create_plan)
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        # Mock mark_step_ordered
-        mock_mark_step_ordered = AsyncMock()
-        monkeypatch.setattr(DcaService, "mark_step_ordered", mock_mark_step_ordered)
-
-        # Mock _place_order_impl to succeed for step 1
-        async def mock_place_order(**kwargs):
-            return {"success": True, "order_id": "ORDER-123"}
-
-        _patch_runtime_attr(monkeypatch, "_place_order_impl", mock_place_order)
-
-        # Mock market data
-        async def mock_sr(_symbol, _market):
-            return {
-                "symbol": "KRW-BTC",
-                "current_price": 100000.0,
-                "supports": [
-                    {"price": 95000.0, "source": "fib_23.6"},
-                ],
-                "resistances": [],
-            }
-
-        async def mock_indicators(_symbol, _indicators, _market):
-            return {
-                "symbol": "KRW-BTC",
-                "price": 100000.0,
-                "indicators": {
-                    "rsi": {"14": 40.0},
-                },
-            }
-
-        _patch_runtime_attr(monkeypatch, "_get_support_resistance_impl", mock_sr)
-        _patch_runtime_attr(monkeypatch, "_get_indicators_impl", mock_indicators)
-
-        # Execute with execute_steps to trigger order placement
-        result = await tools["create_dca_plan"](
-            symbol="KRW-BTC",
-            splits=2,
-            total_amount=100000,
-            strategy="support",
-            dry_run=False,
-            execute_steps=[1],
-        )
-
-        # Verify success and order execution
-        assert result["success"] is True
-        assert result["executed"] is True
-        assert result["plan_id"] == 100
-        assert mock_mark_step_ordered.awaited
-        called_args, _ = mock_mark_step_ordered.await_args
-        assert called_args[0] == 1  # step_id
-        assert called_args[1] == "ORDER-123"  # order_id
-
-    @pytest.mark.asyncio
-    async def test_get_dca_status_impl_monkeypatch(self, monkeypatch):
-        """Test that _get_dca_status_impl can be monkeypatched."""
-        tools = build_tools()
-
-        mock_plan = DcaPlan(
-            id=1,
-            user_id=1,
-            symbol="KRW-ETH",
-            market="crypto",
-            status=DcaPlanStatus.ACTIVE,
-            steps=[],
-        )
-
-        db = AsyncMock()
-        _patch_runtime_attr(
-            monkeypatch,
-            "AsyncSessionLocal",
-            lambda: DummySessionManager(db),
-        )
-
-        async def mock_get_plan(self, _plan_id, _user_id=None):
-            return mock_plan
-
-        monkeypatch.setattr(DcaService, "get_plan", mock_get_plan)
-
-        # Monkeypatch _get_dca_status_impl
-        async def mock_impl(*args, **kwargs):
-            return {
-                "success": True,
-                "plans": [{"plan_id": 999, "symbol": "MONKEY-PATCHED"}],
-                "total_plans": 1,
-            }
-
-        _patch_runtime_attr(monkeypatch, "_get_dca_status_impl", mock_impl)
-
-        # Execute
-        result = await tools["get_dca_status"](plan_id=1)
-
-        # Verify monkeypatched implementation is called
-        assert result["success"] is True
-        assert result["plans"][0]["symbol"] == "MONKEY-PATCHED"
 
 
 class TestPlaceOrderHighAmount:
